wr_ep/tx_header_processor: fix tx_path slave wb interface
If a wb master writes on this slave and it asserts
cyc and stb at the some time, the wb slave in the
tx_path was loosing the first data transaction.
As a result, the frame header was shifted by 16.
The cleanest way to fix this was using the stall line.