wr-pstats-node: redesign of the pstats
The wrsw-pstats was thought to gather statistic of several ports
in a switch. For the wr core is not needed such complexity for
only port. Therefore, the renamed wrsw-pstast, provides the same
functionality but with less complexity and footprint.
The counters implemented are just basic counters, there is room
for improvement.