[streamers/review] renamed xwr_transmission to xwr_streamers and removed gc_
the name of the top entity in the wr_streamers folder was missleading
(xwr_transmission). It was recommended to rename, which I did.
Additionally, names of two modules in wr_streamers start with gc_
which could indicate they are general-cores. This is not the case,
moreover the entity names did not have gc_. I removed the gc_